Migrate generic utils from SDK examples to SDK
