+++ rust/sdk-examples/src/utils/mod.rs
@@ -1,2 +1,1 @@
 pub mod database;
-pub mod util;
