+++ rust/sdk/src/lib.rs
@@ -3,3 +3,4 @@ pub mod steps;
 pub mod test;
 pub mod traits;
 pub mod types;
+pub mod utils;
